+++ inst/scripts/compare_saturation.R
@@ -129,9 +129,9 @@ df <- clean_data(filepath,
 data <- df$data
 codebook <- df$codebook
 
-# Merge codes
-excerpts_merged <- merge_codes(data,
-                               merges = list(
+# Recode themes
+excerpts_recoded <- recode(data,
+                               recodes = list(
                                  c_belonging_connectedness = c(
                                    "c_sense_of_belonging", "c_sense_of_belonging_others", "c_sense_of_belonging_self",
                                    "c_sense_of_connectedness", "c_sense_of_connectedness_family",
@@ -145,11 +145,11 @@ excerpts_merged <- merge_codes(data,
                                  c_suicide_comfort = "Suicide Comfort Conversing"
                                ))
 
-data_merged <- excerpts_merged$data
-codebook_merged <- excerpts_merged$codebook
+data_recode <- excerpts_recoded$data_recode
+codebook_recode <- excerpts_recoded$codebook_recode
 
 # Code summary
-code_summary <- create_code_summary(data_merged, output_type = "tibble",
+code_summary <- create_code_summary(data_recode, output_type = "tibble",
                                     exclude = c("c_priority_excerpt"))
 
 # Compare saturation
@@ -168,5 +168,4 @@ saturation_comparison$results
 
 # View plot
 saturation_comparison$plot
-